intermediate commit during inference debug

+++ floodsens/preprocessing.py
@@ -8,6 +8,7 @@ from floodsens._download import get_copernicus_dem
 from floodsens._process import flow_accumulation, hand, slope, twi
 from floodsens._reproject import reproject_set, reproject_from_raster
 from floodsens.constants import EXTRACT_DICT
+import demloader as dl
 
 
 def _rm_tree(path):
@@ -70,7 +71,9 @@ def extract(zip_path, project_dir, extract_dict):
     return extracted_paths
 
 def download_dem(s2_path, out_dir):
-    dem_path = get_copernicus_dem(s2_path, out_dir)
+    prefixes = dl.prefixes.get_from_raster(s2_path, 30)
+    dem_path = dl.download.from_aws(prefixes, 30, f"{out_dir}/10_DEM.tif")
+    # dem_path = get_copernicus_dem(s2_path, out_dir)
     return dem_path
 
 def clip_dem(dem_path, target_raster_path, project_dir):
@@ -157,6 +160,7 @@ def run_default_preprocessing(project_dir, s2_zip_path, extract_dict=None, delet
     print(f"o---o---o---o---o---o---o\tNot Started \t\t\t(0/7 - {time.time()-mtic:.2f}s|{time.time()-Mtic:.2f}s)")
     if extract_dict is None:
         extract_dict = EXTRACT_DICT
+
     s2_paths_list = extract(s2_zip_path, project_dir, extract_dict)
     target_raster_path = s2_paths_list[0]
     print(f"•---o---o---o---o---o---o\tSentinel bands extracted \t(1/7 - {time.time()-mtic:.2f}s|{time.time()-Mtic:.2f}s)")
@@ -173,7 +177,8 @@ def run_default_preprocessing(project_dir, s2_zip_path, extract_dict=None, delet
     print(f"•---•---•---•---o---o---o\tDEM processed \t\t\t(4/7 - {time.time()-mtic:.2f}s|{time.time()-Mtic:.2f}s)")
     mtic=time.time()
 
-    reprojected_raster_paths = reproject(*all_paths_list, target_raster_path = target_raster_path)
+    # reprojected_raster_paths = reproject(*all_paths_list, target_raster_path = target_raster_path)
+    reprojected_raster_paths = all_paths_list
     print(f"•---•---•---•---•---o---o\tReprojections completed \t(5/7 - {time.time()-mtic:.2f}s|{time.time()-Mtic:.2f}s)")
     mtic=time.time()
 
@@ -285,7 +290,11 @@ def run_multiple_default_preprocessing(project_dir, s2_zip_paths, extract_dict=N
         for stacked_path in stacked_paths:
             Path(stacked_path).unlink()
 
-        Path(merged_path).unlink()
+        if set_type == 'inference': Path(merged_path).unlink()
+        if set_type == 'training':
+            merged_dem_path.unlink()
+            merged_s2_path.unlink()
+            
         print("Unnecessary project files removed.")
